Remove NodeMappingRefreshAction
This exists to execute mapping refresh calls from when elasticsearch
sent dynamic mapping updates from data notes back to the master after
they had indexed the node locally. ES and crate both now apply dynamic
mapping updates before doing any indexing, so this action can be
removed.

Along with the action, we can also now remove the MetadataMappingService,
which was only handling these non-existent refresh calls.
